requirements.txt
python-telegram-bot==21.4
google-generativeai==0.8.2
gspread==6.1.2
oauth2client==4.1.3
python-dotenv==1.0.1
